refactor: review service레이어의 find를 get으로 변경 (#91)
