--- src/main/java/org/example/pedia_777/domain/review/code/ReviewErrorCode.java
@@ -1,16 +1,0 @@
-package org.example.pedia_777.domain.review.code;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.example.pedia_777.common.code.Code;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@RequiredArgsConstructor
-public enum ReviewErrorCode implements Code {
-
-    REVIEW_NOT_FOUND(HttpStatus.NOT_FOUND, "리뷰를 찾을 수 없습니다.");
-
-    private final HttpStatus httpStatus;
-    private final String messageTemplate;
-}
